# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API สำหรับ blog — อ่านได้ทุกคน แก้ได้เฉพาะเจ้าของ",
        "title": "Blog User API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "ล็อกอิน",
                "parameters": [
                    {
                        "description": "email กับ password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_request.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_response.AuthResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "email หรือ password ผิด",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "ดูข้อมูลตัวเอง",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_response.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "สมัครสมาชิก",
                "parameters": [
                    {
                        "description": "ข้อมูลสมัคร",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_request.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_response.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "409": {
                        "description": "email ซ้ำ",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    },
                    "422": {
                        "description": "validate ไม่ผ่าน",
                        "schema": {
                            "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_thitiphongD_blog-user-api_internal_dto_request.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_thitiphongD_blog-user-api_internal_dto_request.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "max=72 เพราะ bcrypt กินได้แค่ 72 bytes เกินกว่านั้นมันตัดทิ้งเงียบๆ",
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8
                }
            }
        },
...
        "github_com_thitiphongD_blog-user-api_internal_dto_response.AuthResponse": {
            "type": "object",
            "properties": {
                "expired_at": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_dto_response.UserResponse"
                }
            }
        },
...
        "github_com_thitiphongD_blog-user-api_internal_dto_response.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_thitiphongD_blog-user-api_internal_response.Body": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_thitiphongD_blog-user-api_internal_response.Pagination"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "github_com_thitiphongD_blog-user-api_internal_response.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "ใส่ \"Bearer \" นำหน้า token ที่ได้จาก /api/v1/auth/login",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}